Move metrics to pkg/shared/metrics
