+++ pkg/cert/legobridge/delegatingprovider.go
@@ -17,9 +17,9 @@ import (
 	"github.com/go-acme/lego/v4/challenge/dns01"
 	"k8s.io/apimachinery/pkg/util/wait"
 
-	"github.com/gardener/cert-management/pkg/cert/metrics"
 	"github.com/gardener/cert-management/pkg/cert/utils"
 	"github.com/gardener/cert-management/pkg/shared"
+	"github.com/gardener/cert-management/pkg/shared/metrics"
 )
 
 // ProviderWithCount is an extended Provider interface.
