+++ pkg/controller/issuer/core/support.go
@@ -25,9 +25,9 @@ import (
 
 	api "github.com/gardener/cert-management/pkg/apis/cert/v1alpha1"
 	"github.com/gardener/cert-management/pkg/cert/legobridge"
-	"github.com/gardener/cert-management/pkg/cert/metrics"
 	"github.com/gardener/cert-management/pkg/cert/utils"
 	ctrl "github.com/gardener/cert-management/pkg/controller"
+	"github.com/gardener/cert-management/pkg/shared/metrics"
 )
 
 // RecoverableError is a recoverable error, i.e. reconcile after same backoff may help
